requirements.txt
Flask
flask-login
mysql-connector-python
argon2-cffi
# requests
# pyotp
# tzlocal
# pytz
